Rename IndexerLogger to IndexerMeta
