+++ runner/src/indexer/indexer.test.ts
@@ -6,7 +6,7 @@ import { VM } from 'vm2';
 import DmlHandler from '../dml-handler/dml-handler';
 import type PgClient from '../pg-client';
 import { type IndexerBehavior } from '../stream-handler/stream-handler';
-import { LogLevel } from '../indexer-logger/indexer-logger';
+import { LogLevel } from '../indexer-meta/indexer-meta';
 
 describe('Indexer unit tests', () => {
   const HASURA_ROLE = 'morgs_near';
