+++ runner/tests/integration.test.ts
@@ -6,11 +6,11 @@ import Indexer from '../src/indexer';
 import HasuraClient from '../src/hasura-client';
 import Provisioner from '../src/provisioner';
 import PgClient from '../src/pg-client';
-import { LogLevel } from '../src/indexer-logger/indexer-logger';
 
 import { HasuraGraphQLContainer, type StartedHasuraGraphQLContainer } from './testcontainers/hasura';
 import { PostgreSqlContainer, type StartedPostgreSqlContainer } from './testcontainers/postgres';
 import block1 from './blocks/00115185108/streamer_message.json';
+import { LogLevel } from '../src/indexer-meta/indexer-meta';
 
 describe('Indexer integration', () => {
   jest.setTimeout(300_000);
